Add `ANONYMOUS_USER_NAME = None` to `base.py`
The creation of the `AnonymousUser` that `django_guardian` did after the `post_migrate` signal ended up failing because there was no `public_group` available yet.
Adding the creation of the `AnonymousUser` from a fixture like our `admin` User is and Reworking the `init.pp` file to run `init_refinery` after migrating, but before the creation of users has addressed this.